+++ src/main/java/pl/edu/uj/kognitywistyka/aboutproject/model/Report.java
@@ -3,14 +3,15 @@ package pl.edu.uj.kognitywistyka.aboutproject.model;
 import java.sql.Date;
 
 public class Report{
-	private int reportId;
+	private long reportId;
 	private String title;
 	private Date data;
 	private String content;
-	public int getReportId() {
+	
+	public long getReportId() {
 		return reportId;
 	}
-	public void setReportId(int raportId) {
+	public void setReportId(long raportId) {
 		this.reportId = raportId;
 	}
 	public String getTitle() {
